requirements.txt
boto3~=1.26.45
python-dotenv==0.21.0